+++ src/button/__tests__/Button.test.tsx
@@ -1,10 +1,15 @@
 import React from 'react';
-import { fireEvent, render, screen, waitFor } from '@testing-library/react';
+import {
+  act,
+  fireEvent,
+  render,
+  screen,
+  waitFor,
+} from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { Button } from '../Button';
 import { BUTTON_TYPE } from '..';
 import userEvent from '@testing-library/user-event';
-import { act } from '@testing-library/react-hooks';
 
 const DummyLoadingButton = ({ loadingLabel }: any) => {
   const [isLoading, setIsLoading] = React.useState(false);
@@ -241,13 +246,5 @@ describe('Button', () => {
     expect(buttons[1].getAttribute('aria-label')).toBe('button-button');
     expect(buttons[2].getAttribute('aria-label')).toBe('submit-button');
     expect(buttons[3].getAttribute('aria-label')).toBe('reset-button');
-  });
-
-  it('should accept tabIndex attribute', () => {
-    render(<Button label="X" tabIndex={1} />);
-
-    const button: any = screen.getByRole('button');
-
-    expect(button.getAttribute('tabindex')).toBe('1');
   });
 });