+++ src/common/components/Hint.tsx
@@ -1,13 +1,14 @@
 import React from 'react';
+import { classNames } from '../utils';
 import { HintProps } from './commonTypes';
 
 export const Hint = ({ text, className, id, useLabel }: HintProps) =>
   useLabel ? (
-    <label id={id} className={className}>
+    <label id={id} className={classNames(['dcx-hint', className])}>
       {text}
     </label>
   ) : (
-    <div id={id} className={className}>
+    <div id={id} className={classNames(['dcx-hint', className])}>
       {text}
     </div>
   );