+++ stories/liveEdit/FormInputLive.tsx
@@ -55,8 +55,7 @@ function FormInputDemo() {
         suffix={{
           content: <></>
         }}
-        staticErrorMessage="static error message"
-        containerClassNameError="govuk-form-group--error"
+        tabIndex={0}
     />
   )
 }